book_favorites_backend/requirements.txt
Django>=5.1.1
mysqlclient
djangorestframework
django-cors-headers
